Changes in download certificates.
start and end are int now.
To download certificates from a CT log server, the job is split into numOfWorkers,
and each one will iteratively make http requests until its portion is completed.
Added UTs.

+++ pkg/mapserver/updater/updater.go
@@ -53,7 +53,7 @@ func NewMapUpdater(root []byte, cacheHeight int) (*MapUpdater, error) {
 
 // UpdateFromCT: download certs from ct log, update the domain entries and update the updates table, and SMT;
 // SMT not committed yet
-func (mapUpdater *MapUpdater) UpdateFromCT(ctx context.Context, ctUrl string, startIdx, endIdx int64) error {
+func (mapUpdater *MapUpdater) UpdateFromCT(ctx context.Context, ctUrl string, startIdx, endIdx int) error {
 	certs, err := logpicker.GetCertMultiThread(ctUrl, startIdx, endIdx, 20)
 	if err != nil {
 		return fmt.Errorf("CollectCerts | GetCertMultiThread | %w", err)
